Add a newline when writing stringified JSONs
While developing a site I've realised that `deno fmt` adds a newline to end of every tracked git file. Since the logic here does not add newlines to end of JSON.strinfigy before adding it to a file, it is not following the format commited to the repository.

This is a small nuisance and since I'm trying things out decided to try and fix it.

Before:
- Clone a repo which uses deco
- Run `deno task start`
- Run `git diff` or `git status` and notice that there's a whitespace removed at the end of `schemas.gen.json`
- Run `deno fmt`
- Run `git diff` or `git status` and notice that such space is gone

After:
- Clone a repo which uses deco
- Run `deno task start`
- Run `git diff` or `git status` and see that there's no change